+++ ft_strncmp.c
@@ -14,9 +14,9 @@
 #include <string.h>
 #include <stdio.h>
 
-int		ft_strncmp(const char *s1, const char *s2, int n)
+int		ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 	
 	i = 0;
 	while (s1[i] == s2[i] && s1[i] != '0' && i < (n - 1))
